wip - refactor with html output working
